% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairycolours.R
\name{fairycolours}
\alias{fairycolours}
\title{Function to extract fairy colours as hex codes}
\usage{
fairycolours(...)
}
\arguments{
\item{...}{Character names of  fairy colours}
}
\description{
Function to extract fairy colours as hex codes
}
