% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairycolours.R
\name{fairypal}
\alias{fairypal}
\title{Return function to interpolate a fairy colour palette}
\usage{
fairypal(palette = "primary", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in fairy_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a fairy colour palette
}
