% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairycolours.R
\name{scale_colour_fairy}
\alias{scale_colour_fairy}
\title{Colour scale for fairy kei inspired coloured palette}
\usage{
scale_colour_fairy(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose a palette by character name in fairypalettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Colour refers to point and line colour on scatter/line plots
}
\examples{
Colour by discrete variable using default palette
ggplot(data, aes(x, y, color = type)) +
 geom_point(size = 4) +
 scale_colour_fairy()
Colour by numeric variable with secondary palette
ggplot(data, aes(x, y, color = type)) +
 geom_point(size = 4, alpha = 0.8) +
 scale_colour_fairy(discrete = FALSE, palette = "secondary")
}
