% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairycolours.R
\name{scale_fill_fairy}
\alias{scale_fill_fairy}
\title{Fill scale for fairy kei inspired coloured palette}
\usage{
scale_fill_fairy(palette = "primary", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Choose a palette by character name in fairypalettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Fill refers to bar or density fill for bar plots
}
\examples{
##Fill by discrete variable with different palette + remove legend (guide)
ggplot(data, aes(x, fill = x)) +
 geom_bar(colour="black") +
 theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
 scale_fill_fairy(palette = "green", guide = "none")
}
